Refactored Bucket files to use ES6 more features - destructuring
